+++ dlfnext-20250310/alibabacloud_dlfnext20250310/models/_create_share_request.py
@@ -8,9 +8,11 @@ class CreateShareRequest(DaraModel):
     def __init__(
         self,
         comment: str = None,
+        enable_write: bool = None,
         share_name: str = None,
     ):
         self.comment = comment
+        self.enable_write = enable_write
         # This parameter is required.
         self.share_name = share_name
 
@@ -25,6 +27,9 @@ class CreateShareRequest(DaraModel):
         if self.comment is not None:
             result['comment'] = self.comment
 
+        if self.enable_write is not None:
+            result['enableWrite'] = self.enable_write
+
         if self.share_name is not None:
             result['shareName'] = self.share_name
 
@@ -34,6 +39,9 @@ class CreateShareRequest(DaraModel):
         m = m or dict()
         if m.get('comment') is not None:
             self.comment = m.get('comment')
+
+        if m.get('enableWrite') is not None:
+            self.enable_write = m.get('enableWrite')
 
         if m.get('shareName') is not None:
             self.share_name = m.get('shareName')