+++ dlfnext-20250310/alibabacloud_dlfnext20250310/models/_share.py
@@ -10,6 +10,7 @@ class Share(DaraModel):
         comment: str = None,
         created_at: int = None,
         created_by: str = None,
+        enable_write: bool = None,
         owner: str = None,
         provider_tenant_id: int = None,
         share_id: str = None,
@@ -20,6 +21,7 @@ class Share(DaraModel):
         self.comment = comment
         self.created_at = created_at
         self.created_by = created_by
+        self.enable_write = enable_write
         self.owner = owner
         self.provider_tenant_id = provider_tenant_id
         self.share_id = share_id
@@ -43,6 +45,9 @@ class Share(DaraModel):
 
         if self.created_by is not None:
             result['createdBy'] = self.created_by
+
+        if self.enable_write is not None:
+            result['enableWrite'] = self.enable_write
 
         if self.owner is not None:
             result['owner'] = self.owner
@@ -74,6 +79,9 @@ class Share(DaraModel):
 
         if m.get('createdBy') is not None:
             self.created_by = m.get('createdBy')
+
+        if m.get('enableWrite') is not None:
+            self.enable_write = m.get('enableWrite')
 
         if m.get('owner') is not None:
             self.owner = m.get('owner')